feat(#34): rename types to @types
