+++ src/database/connection.ts
@@ -1,3 +1,22 @@
-import { createConnections } from 'typeorm';
+import { createConnection, getConnection } from "typeorm";
 
-createConnections();
+const connection = {
+  async create() {
+    await createConnection();
+  },
+
+  async close() {
+    await getConnection().close();
+  },
+
+  async clear() {
+    const connec = getConnection();
+    const entities = connec.entityMetadatas;
+
+    entities.forEach(async (entity) => {
+      const repository = connec.getRepository(entity.name);
+      await repository.query(`DELETE FROM ${entity.tableName}`);
+    });
+  },
+};
+export default connection;
