+++ src/services/AuthenticationUserService.ts
@@ -5,7 +5,7 @@ import { JwtManager } from "../utils/JwtManager";
 import { ResponseHandler } from "../utils/ResponseHandler";
 import User from "../models/User";
 
-import { IRequest, IResponse } from "../types/http";
+import { IRequest, IResponse } from "../@types/http";
 import UserView from "../views/UserView";
 
 export default async function AuthenticationUserService(
