+++ src/utils/JwtManager.ts
@@ -1,7 +1,7 @@
 import "../config/env";
 import jwt from "jsonwebtoken";
 
-import { IDecoded } from "../types/JwtManager";
+import { IDecoded } from "../@types/JwtManager";
 
 export class JwtManager {
   SECRET_KEY: string;
